factory/requirements.txt
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
openpyxl==3.1.2
psycopg2-binary==2.8.6
python-dotenv==0.20.0